// src/event/enums/event.enum.ts
export enum EventMessage {
    NOT_FOUND = 'Event not found',
    ALREADY_EXISTS = 'Event already exists',
    CREATED = 'Event created successfully',
    UPDATED = 'Event updated successfully',
    DELETED = 'Event deleted successfully',
    GET_ALL_EVENTS = 'Get all Events successfully',
    GET_EVENT = 'Get Event successfully',
    TICKET_ALREADY_EXISTS = "Ticket already exists"
}